works in rust and wasm
