+++ src/embed/sync.rs
@@ -1,23 +1,16 @@
 #![allow(clippy::redundant_pattern_matching)] // For derive(DeJson).
 
-use super::http as embed_http;
 use super::types::*;
-use hyper::client::HttpConnector;
-use hyper::Client;
+use crate::fetch;
+use crate::fetch::errors::FetchError;
 use nanoserde::{DeJson, DeJsonErr, SerJson};
 use std::fmt::Debug;
 
-#[cfg(default)]
-pub const USE_BROWSER_FETCH: bool = false;
-
-#[cfg(not(default))]
-pub const USE_BROWSER_FETCH: bool = true;
-
 pub async fn begin_sync(
-    hyper_client: Option<&Client<HttpConnector>>,
+    fetch_client: &fetch::client::Client,
     begin_sync_req: &BeginSyncRequest,
 ) -> Result<BeginSyncResponse, BeginSyncError> {
-    let _pull_resp = pull(hyper_client, begin_sync_req).await?;
+    let _pull_resp = pull(fetch_client, begin_sync_req).await?;
     // TODO do something with the response
     Ok(BeginSyncResponse {})
 }
@@ -62,7 +55,7 @@ pub struct PullResponse {
 
 // client will be none when using browser
 pub async fn pull(
-    hyper_client: Option<&Client<HttpConnector>>,
+    fetch_client: &fetch::client::Client,
     begin_sync_req: &BeginSyncRequest,
 ) -> Result<PullResponse, PullError> {
     let pull_req = PullRequest {
@@ -77,13 +70,7 @@ pub async fn pull(
         &begin_sync_req.diff_server_auth,
     )?;
 
-    let http_resp: http::Response<String>;
-    if USE_BROWSER_FETCH {
-        http_resp = embed_http::browser_fetch(http_req).await?;
-    } else {
-        let client = hyper_client.ok_or_else(|| PullError::NoHttpClient)?;
-        http_resp = embed_http::rust_fetch(client, http_req).await?;
-    }
+    let http_resp: http::Response<String> = fetch_client.request(http_req).await?;
     if http_resp.status() != http::StatusCode::OK {
         return Err(PullError::FetchNotOk(http_resp.status()));
     }
@@ -112,7 +99,7 @@ pub fn new_pull_http_request(
 
 #[derive(Debug)]
 pub enum PullError {
-    FetchFailed(embed_http::FetchError),
+    FetchFailed(FetchError),
     FetchNotOk(http::StatusCode),
     InvalidRequest(http::Error),
     InvalidResponse(DeJsonErr),
@@ -125,8 +112,8 @@ impl From<http::Error> for PullError {
     }
 }
 
-impl From<embed_http::FetchError> for PullError {
-    fn from(err: embed_http::FetchError) -> PullError {
+impl From<FetchError> for PullError {
+    fn from(err: FetchError) -> PullError {
         PullError::FetchFailed(err)
     }
 }
